//session verif

+++ Final_CV/Final_CV/Controllers/ContactsController.cs
@@ -20,60 +20,99 @@ namespace Final_CV.Controllers
         // GET: Contacts
         public ActionResult Index()
         {
-            return View(db.Contacts.ToList());
+            if (HttpContext.Session["LogedUserName"] != null)
+            {
+                return View(db.Contacts.ToList());
+            }
+            else
+            {
+                return RedirectToAction("NotFound", "Home");
+            }
         }
 
         // GET: Contacts/Details/5
         public ActionResult Details(int? id)
         {
-            if (id == null)
+            if (HttpContext.Session["LogedUserName"] != null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                Contact contact = db.Contacts.Find(id);
+                if (contact == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(contact);
             }
-            Contact contact = db.Contacts.Find(id);
-            if (contact == null)
+            else
             {
-                return HttpNotFound();
+                return RedirectToAction("NotFound", "Home");
             }
-            return View(contact);
-        }
+            }
 
         // GET: Contacts/Create
         public ActionResult Create()
         {
-            return View();
+            if (HttpContext.Session["LogedUserName"] != null)
+              {
+                return View();
+              }
+            else
+            {
+                return RedirectToAction("NotFound", "Home");
+             }
         }
 
-        // POST: Contacts/Create
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
+// POST: Contacts/Create
+// To protect from overposting attacks, please enable the specific properties you want to bind to, for 
+// more details see http://go.microsoft.com/fwlink/?LinkId=317598.
+[HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "ContactID,Label,value,Logo")] Contact contact)
         {
-            if (ModelState.IsValid)
+            if (HttpContext.Session["LogedUserName"] != null)
             {
-                db.Contacts.Add(contact);
-                db.SaveChanges();
-                return RedirectToAction("Index");
-            }
 
-            return View(contact);
+                if (ModelState.IsValid)
+                    {
+                        db.Contacts.Add(contact);
+                        db.SaveChanges();
+                        return RedirectToAction("Index");
+                    }
+                 return View(contact);
+                }
+                else
+                {
+                    return RedirectToAction("NotFound", "Home");
+                }
         }
 
         // GET: Contacts/Edit/5
         public ActionResult Edit(int? id)
         {
-            if (id == null)
+            if (HttpContext.Session["LogedUserName"] != null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+
+                if (id == null)
+                    {
+                        return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                    }
+                    Contact contact = db.Contacts.Find(id);
+                    if (contact == null)
+                    {
+                        return HttpNotFound();
+                    }
+                    return View(contact);
             }
-            Contact contact = db.Contacts.Find(id);
-            if (contact == null)
+            else
             {
-                return HttpNotFound();
+                return RedirectToAction("NotFound", "Home");
             }
-            return View(contact);
+
+
         }
 
         // POST: Contacts/Edit/5
@@ -83,28 +122,48 @@ namespace Final_CV.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "ContactID,Label,value,Logo")] Contact contact)
         {
-            if (ModelState.IsValid)
+
+            if (HttpContext.Session["LogedUserName"] != null)
             {
-                db.Entry(contact).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Index");
+
+
+                if (ModelState.IsValid)
+                    {
+                        db.Entry(contact).State = EntityState.Modified;
+                        db.SaveChanges();
+                        return RedirectToAction("Index");
+                    }
+                 return View(contact);
             }
-            return View(contact);
+            else
+            {
+                return RedirectToAction("NotFound", "Home");
+            }
+
         }
 
         // GET: Contacts/Delete/5
         public ActionResult Delete(int? id)
         {
-            if (id == null)
+            if (HttpContext.Session["LogedUserName"] != null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+
+
+                if (id == null)
+                    {
+                        return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                    }
+                    Contact contact = db.Contacts.Find(id);
+                    if (contact == null)
+                    {
+                        return HttpNotFound();
+                    }
+                    return View(contact);
             }
-            Contact contact = db.Contacts.Find(id);
-            if (contact == null)
+            else
             {
-                return HttpNotFound();
+                return RedirectToAction("NotFound", "Home");
             }
-            return View(contact);
         }
 
         // POST: Contacts/Delete/5
@@ -112,10 +171,18 @@ namespace Final_CV.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
-            Contact contact = db.Contacts.Find(id);
-            db.Contacts.Remove(contact);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            if (HttpContext.Session["LogedUserName"] != null)
+            {
+
+                Contact contact = db.Contacts.Find(id);
+                db.Contacts.Remove(contact);
+                db.SaveChanges();
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                return RedirectToAction("NotFound", "Home");
+            }
         }
 
         protected override void Dispose(bool disposing)
